% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Accelerated stability data}
\format{
An object of class \code{"data.frame"}
\describe{
\item{Celsius}{The temperature in celsius.}
\item{time}{Times for which the datapoints are gathered.}
\item{K}{The temperature in Kelvin.}
\item{true}{}
\item{conc}{The concentration at a time.}
}
}
\usage{
data(example_data)
}
\description{
Small, artificially generated toy data set that contains accelerated stability
data for the purpose of the function examples.
}
\examples{
data(example_data)
head(example_data)


}
\references{
This data set was artificially created for the package.
}
\keyword{dataset}
