% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potency.R
\docType{data}
\name{potency}
\alias{potency}
\title{Potency Accelerated Stability Data}
\format{
An object of class \code{"data.frame"} with 78 rows and 3 variables
\describe{
\item{Time}{Time for which the datapoints are gathered.}
\item{Potency}{Measured potency at a time.}
\item{Celsius}{The temperature in celsius.}

}
}
\usage{
data(potency)
}
\description{
An example dataset containing potency data at different
temperatures..
}
\keyword{dataset}
