% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_limit.R
\name{release_limit}
\alias{release_limit}
\title{Release Limit}
\usage{
release_limit(
  step1_down_object,
  shelf_temp,
  shelf_time,
  LSL,
  interval = "CI",
  confidence_interval = 0.95,
  xname = NULL,
  yname = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{step1_down_object}{The fit object from the step1.down function (required).}

\item{shelf_temp}{The temperature at which the product will be stored.}

\item{shelf_time}{The time for which the product will be stored.}

\item{LSL}{A numeric value for the lower specification.}

\item{interval}{Choose between confidence intervals or prediction intervals (default is CI).}

\item{confidence_interval}{Confidence level for the confidence and prediction intervals
around the predictions (default 0.95).}

\item{xname}{Label for the x-axis (optional).}

\item{yname}{Label for the y-axis (optional).}

\item{xlim}{x-axis limits (optional).}

\item{ylim}{y-axis limits (optional).}
}
\value{
Returns release limit calculations and plot.
}
\description{
Calculate the Release Limit for a product.
}
\details{
Use the output from step1.down to run a release limit calculation.
}
\examples{
#load antigenicity data
data(antigenicity)

#run step1.down fit
fit1 <- step1_down(data = antigenicity, y = "conc", .time = "time",
 C = "Celsius",max_time_pred = 3)

release_lim <- release_limit(step1_down_object = fit1, shelf_temp = 5, shelf_time = 3,
                             LSL = 45, interval = "CI", confidence_interval = 0.95)
release_lim$RL_plot

}
