% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenicity.R
\docType{data}
\name{antigenicity}
\alias{antigenicity}
\title{Antigenicity Accelerated Stability Data}
\format{
An object of class \code{"data.frame"} with 50 rows and 5 variables
\describe{
\item{time}{Number of days in years for which the datapoints are gathered.}
\item{conc}{The concentration at a time.}
\item{K}{The temperature in Kelvin.}
\item{Celsius}{The temperature in celsius.}
\item{days}{Number of days for which the datapoints are gathered.}

}
}
\usage{
data(antigenicity)
}
\description{
An example dataset containing antigenicity concentration data at different
temperatures over a period of up to 147 days.
}
\keyword{dataset}
